+++ database/migrations/20180817064720_add_replace_id_to_messages_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddReplaceIdToMessagesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('messages', function(Blueprint $table) {
+            $table->string('replaceid', 64)->nullable();
+            $table->index('replaceid');
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('infos', function(Blueprint $table) {
+            $table->dropColumn('replaceid');
+            $table->dropIndex('messages_replaceid_index');
+        });
+    }
+}